+++ fe/src/app.tsx
@@ -1,11 +1,7 @@
-import styled from 'styled-components';
+import { WeatherImage, WeatherText } from '@/components';
+import { styled } from '@/utils';
 
-import sunnyImage from '@/assets/images/sunny.jpg';
-
-const WeatherBackground = styled.div`
-  width: 100vw;
-  height: 100vh;
-`;
+/**********************************************************************************/
 
 const WeatherStyle = styled.div`
   display: flex;
@@ -13,36 +9,13 @@ const WeatherStyle = styled.div`
   gap: 2em;
 `;
 
-const WeatherImageStyle = styled.div`
-  min-height: 50vh;
-  background-image: url(${sunnyImage});
-  background-repeat: no-repeat;
-  background-size: contain;
-`;
-
-const WeatherText = styled.div`
-  margin-top: 1.33em;
-  font-size: 1.66em;
-  text-align: center;
-`;
-
 /**********************************************************************************/
-
-function WeatherImage() {
-  return <WeatherImageStyle></WeatherImageStyle>;
-}
-
-function WeatherInfo() {
-  return <WeatherText>Bamba</WeatherText>;
-}
 
 export default function App() {
   return (
-    <WeatherBackground>
-      <WeatherStyle>
-        <WeatherImage></WeatherImage>
-        <WeatherInfo></WeatherInfo>
-      </WeatherStyle>
-    </WeatherBackground>
+    <WeatherStyle>
+      <WeatherImage></WeatherImage>
+      <WeatherText></WeatherText>
+    </WeatherStyle>
   );
 }
