+++ fe/src/assets/styles/index.css
@@ -16,57 +16,23 @@
 /**********************************************************************************/
 
 :root {
-  font-family: 'Poppins';
-  line-height: 1.33;
-  font-weight: bolder;
-
   color-scheme: light dark;
   color: rgba(255, 255, 255, 0.87);
-
+  font-family: 'Poppins';
   font-synthesis: none;
   /* Allows kerning and optional ligatures */
   /* See: https://www.gnu.org/software/groff/manual/groff.html.node/Ligatures-and-Kerning.html */
   text-rendering: optimizeLegibility;
 }
 
-a {
-  font-weight: 500;
-  color: #646cff;
-  /* Inherit text decoration options from the the parent element */
-  text-decoration: inherit;
-}
-a:hover {
-  color: #535bf2;
+#root {
+  width: 100vw;
+  height: 100vh;
 }
 
 body {
   /* Removes the default margin supplied by the browser */
   margin: 0 auto;
-}
-
-h1 {
-  /* The em unit means "my parent element's font-size" in the case of typography */
-  /* See: https://developer.mozilla.org/en-US/docs/Web/CSS/font-size#ems */
-  font-size: 3.2em;
-  line-height: 1.1;
-}
-
-button {
-  /* The rem unit means "The root element's font-size" in the case of typography */
-  /* See: https://developer.mozilla.org/en-US/docs/Web/CSS/font-size#rems */
-  height: 2.5rem;
-  width: 9rem;
-  border-radius: 30px;
-  border-color: transparent;
-  /* outline is the line drawn around an element (outside the border) */
-  outline-style: none;
-  color: white;
-  font-weight: bolder;
-  /* The default size is 16px, we want to make sure (across browsers) that the */
-  /* font size of the button matches that size (some browser have different font */
-  /* size for browsers) */
-  font-size: 1rem;
-  cursor: pointer;
 }
 
 @media (prefers-color-scheme: light) {
@@ -80,14 +46,4 @@ button {
   button {
     background-color: #f9f9f9;
   }
-}
-
-*::-webkit-scrollbar {
-  width: 8px;
-  background-color: white;
-}
-
-*::-webkit-scrollbar-thumb {
-  border-radius: 5px;
-  background-color: rgba(114, 130, 214, 0.46);
 }
