+++ fe/src/components/index.ts
@@ -1,0 +1,6 @@
+import WeatherImage from './weather-image.tsx';
+import WeatherText from './weather-text.tsx';
+
+/**********************************************************************************/
+
+export { WeatherImage, WeatherText };
