+++ fe/src/components/weather-text.tsx
@@ -1,0 +1,15 @@
+import { styled } from '@/utils';
+
+/**********************************************************************************/
+
+const WeatherTextStyle = styled.div`
+  margin-top: 1.33em;
+  font-size: 1.66em;
+  text-align: center;
+`;
+
+/**********************************************************************************/
+
+export default function WeatherText() {
+  return <WeatherTextStyle>Placeholder</WeatherTextStyle>;
+}
